… Project.toml
… src/newton.jl
… src/parameters.jl
… src/types.jl
… test/bene_testAmica.jl
… test/compare_amica_implementations.jl
… test/runtests.jl