@behinger 2023-12-19
for a large problem I get ~12s/iteration in julia vs. 3s in fortran. So speedup of fortran is still factor 4 or so